Move Simple History customizations on their own module as well as settings
