+++ conf/__init__.py
@@ -11,3 +11,5 @@ from .apps import *
 from .s3 import *
 
 from .iot import *
+
+from .simple_history import *
